# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "CambioDePerfil": {
        "properties": {
          "correo": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correo"
          },
          "nombre": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          }
        },
        "title": "CambioDePerfil",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MiCita": {
        "description": "Una cita vista **por quien la reservó**: dónde, cuándo y con qué servicios.",
        "properties": {
          "estado": {
            "$ref": "#/components/schemas/EstadoReserva"
          },
          "fin": {
            "format": "date-time",
            "title": "Fin",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "inicio": {
            "format": "date-time",
            "title": "Inicio",
            "type": "string"
          },
          "negocio": {
            "title": "Negocio",
            "type": "string"
          },
          "negocio_slug": {
            "title": "Negocio Slug",
            "type": "string"
          },
          "se_puede_cancelar": {
            "title": "Se Puede Cancelar",
            "type": "boolean"
          },
          "servicios": {
            "items": {
              "$ref": "#/components/schemas/ServicioDeMiCita"
            },
            "title": "Servicios",
            "type": "array"
          },
          "total_centavos": {
            "title": "Total Centavos",
            "type": "integer"
          },
          "zona_horaria": {
            "title": "Zona Horaria",
            "type": "string"
          }
        },
        "required": [
          "id",
          "negocio",
          "negocio_slug",
          "inicio",
          "fin",
          "estado",
          "zona_horaria",
          "servicios",
          "total_centavos",
          "se_puede_cancelar"
        ],
        "title": "MiCita",
        "type": "object"
      },
...
      "MiPerfil": {
        "description": "Lo que la persona ve de sí misma. El teléfono sí, porque es suyo.",
        "properties": {
          "correo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correo"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "idioma": {
            "title": "Idioma",
            "type": "string"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          },
          "telefono": {
            "title": "Telefono",
            "type": "string"
          },
          "telefono_verificado": {
            "title": "Telefono Verificado",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "nombre",
          "telefono",
          "telefono_verificado",
          "correo",
          "idioma"
        ],
        "title": "MiPerfil",
        "type": "object"
      },
...
      "NegocioFavorito": {
        "description": "Un salón guardado. Misma forma que un resultado de búsqueda, para pintarlo igual.",
        "properties": {
          "abierto_ahora": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abierto Ahora"
          },
          "categorias": {
            "items": {
              "type": "string"
            },
            "title": "Categorias",
            "type": "array"
          },
          "direccion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direccion"
          },
          "foto_portada": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foto Portada"
          },
          "negocio_id": {
            "format": "uuid",
            "title": "Negocio Id",
            "type": "string"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          },
          "numero_reviews": {
            "title": "Numero Reviews",
            "type": "integer"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "servicios_desde_centavos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Servicios Desde Centavos"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "zona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zona"
          }
        },
        "required": [
          "negocio_id",
          "slug",
          "nombre",
          "direccion",
          "zona",
          "foto_portada",
          "rating",
          "numero_reviews",
          "servicios_desde_centavos",
          "categorias",
          "abierto_ahora"
        ],
        "title": "NegocioFavorito",
        "type": "object"
      },
...
      "NuevoFavorito": {
        "properties": {
          "negocio_id": {
            "format": "uuid",
            "title": "Negocio Id",
            "type": "string"
          }
        },
        "required": [
          "negocio_id"
        ],
        "title": "NuevoFavorito",
        "type": "object"
      },
...
      "PeticionDeReserva": {
        "properties": {
          "inicio": {
            "format": "date-time",
            "title": "Inicio",
            "type": "string"
          },
          "negocio_slug": {
            "title": "Negocio Slug",
            "type": "string"
          },
          "nombre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre"
          },
          "nota": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nota"
          },
          "profesional_id": {
            "format": "uuid",
            "title": "Profesional Id",
            "type": "string"
          },
          "servicios": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "minItems": 1,
            "title": "Servicios",
            "type": "array"
          }
        },
        "required": [
          "negocio_slug",
          "servicios",
          "inicio",
          "profesional_id"
        ],
        "title": "PeticionDeReserva",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Reservas y marketplace de belleza y bienestar en Panamá. Cada endpoint cita el requisito del brief que cubre.",
    "title": "API de Bukeo",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/mi/favoritos": {
      "get": {
        "description": "Los guardados, en orden de guardado descendente: el último arriba.",
        "operationId": "listar_favoritos_api_v1_mi_favoritos_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NegocioFavorito"
                  },
                  "title": "Response Listar Favoritos Api V1 Mi Favoritos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mis salones guardados (MKT-5)",
        "tags": [
          "cliente"
        ]
      },
      "post": {
        "description": "Guardar dos veces el mismo salón **no es un error**: es la misma respuesta.\n\nEl botón de favorito se pulsa con el dedo en un teléfono con 3G, y un doble toque que\ndevuelve `409` es una pantalla en rojo por hacer bien lo que se quería hacer. Se resuelve\ncon `ON CONFLICT DO NOTHING`, que además es a prueba de dos peticiones a la vez.",
        "operationId": "guardar_favorito_api_v1_mi_favoritos_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NuevoFavorito"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Guardar Favorito Api V1 Mi Favoritos Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Guardar un salón (MKT-5)",
        "tags": [
          "cliente"
        ]
      }
    },
    "/api/v1/mi/favoritos/{negocio_id}": {
      "delete": {
        "description": "Quitar algo que no estaba tampoco es un error: el resultado que se pedía ya se cumple.",
        "operationId": "quitar_favorito_api_v1_mi_favoritos__negocio_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "negocio_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Negocio Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Quitar de favoritos (MKT-5)",
        "tags": [
          "cliente"
        ]
      }
    },
...
    "/api/v1/mi/perfil": {
      "get": {
        "operationId": "leer_perfil_api_v1_mi_perfil_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiPerfil"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mis datos",
        "tags": [
          "cliente"
        ]
      },
      "patch": {
        "description": "El **teléfono no se cambia por aquí**: es la identidad de la cuenta.\n\nCambiarlo es verificar el nuevo con un OTP, porque si bastara con un `PATCH`, quien\nentrara una vez en una sesión ajena se llevaría la cuenta entera cambiando el número.",
        "operationId": "editar_perfil_api_v1_mi_perfil_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambioDePerfil"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiPerfil"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cambiar mi nombre o mi correo",
        "tags": [
          "cliente"
        ]
      }
    },
    "/api/v1/mi/reservas": {
      "get": {
        "description": "El historial de la persona, en todos los salones donde ha reservado.\n\nSe apoya en `client_user_id`, que está desnormalizado a propósito: si hubiera que cruzar\nlas fichas de cliente de cada negocio, la pantalla de inicio de la app haría una consulta\npor salón (ADR sobre el modelo, §8.1).",
        "operationId": "mis_reservas_api_v1_mi_reservas_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MiCita"
                  },
                  "title": "Response Mis Reservas Api V1 Mi Reservas Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mis reservas (RSV-7)",
        "tags": [
          "cliente"
        ]
      },
      "post": {
        "description": "Crea la cita. Devuelve `409 SLOT_NO_DISPONIBLE` si alguien se adelantó.\n\nMirar un hueco no lo aparta: se compite por él al confirmar, y quien pierde recibe un\nmensaje que entiende en vez de una cita a otra hora que no eligió.",
        "operationId": "reservar_api_v1_mi_reservas_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeticionDeReserva"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiCita"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reservar (RSV-1, AGD-4)",
        "tags": [
          "cliente"
        ]
      }
    },
...
    "/api/v1/mi/reservas/{reserva_id}/repetir": {
      "get": {
        "description": "Devuelve el mismo servicio y el mismo profesional, listos para el flujo de reserva.\n\n**No crea nada.** «Reservar de nuevo» es rellenar el formulario, no saltárselo: la hora\nsigue eligiéndose, porque el hueco de hace un mes no significa nada hoy y porque una cita\ncreada sin que nadie confirme la hora es una sorpresa el día que toca.",
        "operationId": "repetir_reserva_api_v1_mi_reservas__reserva_id__repetir_get",
        "parameters": [
          {
            "in": "path",
            "name": "reserva_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reserva Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservaDeNuevo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reservar de nuevo (MKT-5)",
        "tags": [
          "cliente"
        ]
      }
    },
...
  }
}